Dbms implmnt (#5)
* Finished Basic Database implementation

* Finished Basic Database implementation
Excluded facebook friend syncing for the time being, since there is some confusion in Graph API's working. Currently the Driver activity and Passenger activity has only placeholder text.


* Testing phase dbms_implementation
Fixed most of crashes, program correctly inserts to database the login details and User details.
To implement insertion of friends to users node

* Testing Midway dbms_implementation
There a few known errors in this version, and some crashes, especially when accessing the database.



~* Mid-way Implement Database

This commit includes or tries to include only the changes related to Database only and doesn't include map implementation codes. Hopefully this commit will be safe to revert in case of any errors~