Implemented Driver Detail selection activity
In this activity the driver enters the details of car pool like location from and to, the time of carpool,the date of car pool and the number of seats in the car. Also implemented checks to ensure that the time and date are valid and so is the number of seat. This commit contains a few peices of code from the next iteration, which is the implementaion of list of requests from the passengers who want to car pool with the driver. All the code in .java files have been commented out to prevent confusion and error. The .xml files and model file MyRequest.java have been added as is.