Added Recycler view for list of passengers requesting. Ability to Accept & Reject by Driver. NOT TESTED yet.
Added the recycler view which holds the list of all passengers who are requesting carpool from a driver. Currently this will always be zero or none as the passenger logic is not finished yet. The recycler view houses the name of the passenger, the location where they like to picked up from and 3 buttons to 1. see the place on the map,2. accept the passenger,3. reject the passenger. The changes made in this quick iteration has not been tested and/or it is expected to fail in some way