Fixed Driver's list of Requestees(recycleView) not rendering reliably
The list of passengers requesting a Driver to be included in car pool will now be rendered reliably, the show on map button has been fixed to show the correct location in a zoomed in state instead of randomly adding marker somewhere in the corner. Accept and Reject buttons are still not implemented yet. Work on passenger's side of view has begun with a few basic things added and none of it are tested. Also this commit includes the changes made by @TrueEntropy in the colors.xml file.