Implemented Friend system for the driver request purpose
User's can now manually add friends. The option to add friends is accessed by pressing the 3 dots in the corner and selecting 'Add friend'. Option to Logout is available, so the user is not trapped until the carpool is over. It is also accessed from the same 3 dots in the corner, this obviously will not be available in the login activity. Replaced file FriendsModel.java with FriendReqModel.java.